Further separated RIP from Django-RIP
