+++ django_adapter/view/view_resource.py
@@ -1,3 +1,7 @@
+from rip.view import view_pipeline_factory
+from rip.view.view_actions import ViewActions
+
+from django_adapter.view.decorators import validate_view_action
 from rip.generic_steps.default_authentication import \
     DefaultAuthentication
 from rip.generic_steps.default_data_cleaner import DefaultRequestCleaner
@@ -11,9 +15,6 @@ from rip.generic_steps.default_schema_validation import \
     DefaultSchemaValidation
 from rip.generic_steps.default_view_actions import DefaultViewActions
 from rip.generic_steps.default_view_authorization import DefaultViewAuthorization
-from rip.view import view_pipeline_factory
-from rip.view.decorators import validate_view_action
-from rip.view.view_actions import ViewActions
 
 
 class ViewResource(object):
