+++ rip/generic_steps/default_entity_actions.py
@@ -1,6 +1,5 @@
-from django.conf import settings
+from rip.generic_steps import error_types
 from rip.response import Response
-from rip import error_types
 
 
 class DefaultEntityActions(object):
