+++ rip/generic_steps/default_schema_validation.py
@@ -1,5 +1,5 @@
-from rip import error_types
 from rip.crud.crud_actions import CrudActions
+from rip.generic_steps import error_types
 from rip.response import Response
 from rip.schema.default_field_value import \
     DEFAULT_FIELD_VALUE
