+++ rip/tests/unit_tests/generic_steps/test_default_authentication.py
@@ -4,9 +4,9 @@ from hamcrest.core import assert_that
 from hamcrest.core.core.isequal import equal_to
 from mock import MagicMock
 
-from rip.generic_steps import default_authentication
-from rip.request import Request
 from rip import error_types
+from rip.generic_steps import default_authentication, error_types
+from rip.request import Request
 
 
 class TestAuthenticationStep(unittest.TestCase):
