+++ rip/tests/unit_tests/generic_steps/test_default_schema_validation.py
@@ -5,11 +5,11 @@ from hamcrest.core.core.isequal import equal_to
 from hamcrest.library.collection.issequence_containing import has_item, \
     has_items
 
-from rip import error_types
-from rip.api_schema import ApiSchema
 from rip.crud.crud_actions import CrudActions
+from rip.generic_steps import error_types
 from rip.generic_steps.default_schema_validation import \
     DefaultSchemaValidation
+from rip.schema.api_schema import ApiSchema
 from rip.schema.boolean_field import \
     BooleanField
 from rip.schema.schema_field import SchemaField
