+++ rip/tests/unit_tests/schema/test_choice_field.py
