+++ rip/tests/unit_tests/schema/test_list_field.py
