+++ rip/tests/unit_tests/schema/test_resource_uri_field.py
