+++ tests/integration_tests/blank_test_resource.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
 from django_adapter.crud_resource import CrudResource
-from rip.api_schema import ApiSchema
 from rip.crud.crud_actions import CrudActions
 from rip.generic_steps.default_entity_actions import DefaultEntityActions
+from rip.schema.api_schema import ApiSchema
 from rip.schema.string_field import StringField
