+++ tests/integration_tests/test_delete_detail.py
@@ -1,7 +1,7 @@
 from hamcrest import assert_that, equal_to
 from mock import patch
 
-from rip import error_types
+from rip.generic_steps import error_types
 from tests import request_factory
 from tests.integration_tests.person_base_test_case import \
     PersonResourceBaseTestCase
