+++ tests/integration_tests/test_get_list.py
@@ -1,6 +1,6 @@
 from hamcrest import assert_that, equal_to
 
-from rip import error_types
+from rip.generic_steps import error_types
 from tests import request_factory
 from tests.integration_tests.person_base_test_case import \
     PersonResourceBaseTestCase
