+++ tests/unit_tests/django_adapter/test_crud_resource.py
@@ -5,11 +5,11 @@ from mock import MagicMock, patch
 from django_adapter.crud_resource import CrudResource, CrudActions, \
     crud_pipeline_factory
 from rip import error_types
-from rip.api_schema import ApiSchema
-from rip.generic_steps import default_authentication
+from rip.crud.pipeline_composer import PipelineComposer
+from rip.generic_steps import default_authentication, error_types
 from rip.generic_steps.default_entity_actions import \
     DefaultEntityActions
-from rip.pipeline_composer import PipelineComposer
+from rip.schema.api_schema import ApiSchema
 from rip.schema.boolean_field import \
     BooleanField
 from rip.schema.string_field import StringField
